Move db to dal
