+++ data-access-layer/db/JobRunDetailSchema.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const Mongoose = require('mongoose');
-const CONST = require('../../../common/constants');
+const CONST = require('../../common/constants');
 
 const JobRunDetailSchema = new Mongoose.Schema({
   name: {
